src/requirements.txt
streamlit
streamlit-chat
streamlit-extras
langchain
langchain-core
langchain-openai
langchain_community
langchain-chroma
python-dotenv
crewai
crewai[tools]
embedchain[postgres]
browserbase
pydantic
chromadb
tiktoken
PyPDF2
python-docx
llama_index
llama-index-embeddings-langchain
llama-index-vector-stores-faiss
transformers
torch
huggingface-hub
pandas
faiss-cpu
FPDF
